requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy[asyncio]>=2.0
asyncpg
psycopg2-binary
alembic
python-dotenv
pydantic-settings
minio
python-multipart
uuid
httpx
aiocache
Pillow
boto3==1.35.99
aiosqlite
requests
diskcache
pytest
pytest-asyncio
swagger-ui-bundle
podman-compose